Refactor into optimize_sdfield library
